add invites panel
