+++ packages/frontend/src/components/UserSettingsModal/OrganizationPanel.tsx
@@ -1,22 +1,9 @@
 import React, { FC, useEffect, useState } from 'react';
 import { useMutation, useQueryClient } from 'react-query';
-import { ApiError, formatTimestamp } from 'common';
-import {
-    Button,
-    FormGroup,
-    InputGroup,
-    Intent,
-    Callout,
-} from '@blueprintjs/core';
-import CopyToClipboard from 'react-copy-to-clipboard';
+import { ApiError } from 'common';
+import { Button, FormGroup, InputGroup, Intent } from '@blueprintjs/core';
 import { useApp } from '../../providers/AppProvider';
 import { lightdashApi } from '../../api';
-import {
-    useCreateInviteLinkMutation,
-    useRevokeInvitesMutation,
-} from '../../hooks/useInviteLink';
-import { useTracking } from '../../providers/TrackingProvider';
-import { EventName } from '../../types/Events';
 
 const updateOrgQuery = async (data: { organizationName: string }) =>
     lightdashApi<undefined>({
@@ -33,12 +20,9 @@ const OrganizationPanel: FC = () => {
         showToastSuccess,
         user,
     } = useApp();
-    const { track } = useTracking();
     const [organizationName, setOrganizationName] = useState<
         string | undefined
     >(user.data?.organizationName);
-    const inviteLink = useCreateInviteLinkMutation();
-    const revokeInvitesMutation = useRevokeInvitesMutation();
     const { isLoading, error, mutate } = useMutation<
         undefined,
         ApiError,
@@ -95,74 +79,6 @@ const OrganizationPanel: FC = () => {
                     onChange={(e) => setOrganizationName(e.target.value)}
                 />
             </FormGroup>
-            <div>
-                <FormGroup
-                    label="Invite users to your organization"
-                    labelFor="invite-link-input"
-                >
-                    {inviteLink.data ? (
-                        <>
-                            <InputGroup
-                                id="invite-link-input"
-                                type="text"
-                                readOnly
-                                value={`${window.location.protocol}//${window.location.host}/invite/${inviteLink.data.inviteCode}`}
-                                rightElement={
-                                    <CopyToClipboard
-                                        text={`${window.location.protocol}//${window.location.host}/invite/${inviteLink.data.inviteCode}`}
-                                        options={{ message: 'Copied' }}
-                                        onCopy={() =>
-                                            showToastSuccess({
-                                                title: 'Invite link copied',
-                                            })
-                                        }
-                                    >
-                                        <Button minimal icon="clipboard" />
-                                    </CopyToClipboard>
-                                }
-                            />
-                            <Callout intent="primary" style={{ marginTop: 10 }}>
-                                Share this link with your colleagues and they
-                                can join your organization. This link will
-                                expire at{' '}
-                                <b>
-                                    {formatTimestamp(inviteLink.data.expiresAt)}
-                                </b>
-                            </Callout>
-                        </>
-                    ) : (
-                        <Button
-                            text="Create invite link"
-                            style={{ marginTop: 10 }}
-                            loading={inviteLink.isLoading}
-                            onClick={() => {
-                                track({
-                                    name: EventName.INVITE_BUTTON_CLICKED,
-                                });
-                                inviteLink.mutate();
-                            }}
-                        />
-                    )}
-                </FormGroup>
-                <Callout intent="warning" style={{ marginTop: 20 }}>
-                    <p>
-                        This action will remove all active and expired invites
-                        links.
-                    </p>
-                    <Button
-                        intent="danger"
-                        text="Revoke all invites"
-                        loading={inviteLink.isLoading}
-                        onClick={() => {
-                            track({
-                                name: EventName.REVOKE_INVITES_BUTTON_CLICKED,
-                            });
-                            revokeInvitesMutation.mutate();
-                            inviteLink.reset();
-                        }}
-                    />
-                </Callout>
-            </div>
             <div style={{ flex: 1 }} />
             <Button
                 style={{ alignSelf: 'flex-end', marginTop: 20 }}
